requirements.txt
Django==3.2
psycopg2>=2.8
Babel
django-phonenumber-field[phonenumberslite]
reportlab